financeCopilot/agents/requirements.txt
# Google Gemini için gerekli paket
google-generativeai>=0.4.0

# JSON işleme ve veri temizliği
orjson>=3.9.7

# Ortam değişkenlerini yönetmek için
python-dotenv>=1.0.1

# JSON verisini güzel yazdırmak için (isteğe bağlı)
rich>=13.7.0

# OCR ve Görüntü İşleme
pillow
pytesseract

# PDF İşleme
PyMuPDF

# Ortak kullanım için
requests

# Testler için
gradio

# PDF'den metin çıkarmak için
pdfplumber

# Backend iletişim için
fastapi

# Asenkron HTTP istemcisi
httpx

